Util classes and functions to reconstruct the network hierarchy from a data plane proto.
The networks, instances, subnets given in the protobuf are flattened and this commit contains code to reconstruct the hierarchy
e.g., putting the subnets under regions and projects.

The classes defined here are only a hierarchical index for easier access to the network topology, and do not contain any
dataplane states, e.g. if the network peer is active, and if the peer enables importing/exporting custom routes.

Full hierarchy:
    - Internet > Project > VPC Network <> Region
    - Region > Subnet > IP Range
    - Region > VM > NIC
    - Region > Cloud Router, VPN gateway